Refactor YAML keywords and tag completion out of config namespace
