+++ src/main/java/fr/adrienbrault/idea/symfony2plugin/util/completion/YamlKeywordsCompletionProvider.java
@@ -1,4 +1,4 @@
-package fr.adrienbrault.idea.symfony2plugin.util.yaml;
+package fr.adrienbrault.idea.symfony2plugin.util.completion;
 
 import com.intellij.codeInsight.completion.CompletionParameters;
 import com.intellij.codeInsight.completion.CompletionProvider;
@@ -11,6 +11,7 @@ import com.intellij.psi.impl.source.tree.LeafPsiElement;
 import com.intellij.util.ProcessingContext;
 import com.jetbrains.jsonSchema.ide.JsonSchemaService;
 import com.jetbrains.jsonSchema.impl.JsonSchemaObject;
+import fr.adrienbrault.idea.symfony2plugin.util.yaml.YamlHelper;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.HashMap;
