+++ src/main/java/fr/adrienbrault/idea/symfony2plugin/util/completion/YamlTagCompletionProvider.java
@@ -1,4 +1,4 @@
-package fr.adrienbrault.idea.symfony2plugin.util.yaml;
+package fr.adrienbrault.idea.symfony2plugin.util.completion;
 
 import com.intellij.codeInsight.completion.*;
 import com.intellij.codeInsight.lookup.LookupElement;
@@ -13,6 +13,7 @@ import com.intellij.psi.util.PsiTreeUtil;
 import com.intellij.util.ProcessingContext;
 import fr.adrienbrault.idea.symfony2plugin.Symfony2ProjectComponent;
 import fr.adrienbrault.idea.symfony2plugin.util.SymfonyUtil;
+import fr.adrienbrault.idea.symfony2plugin.util.yaml.YamlHelper;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.yaml.YAMLTokenTypes;
