+++ src/test/java/fr/adrienbrault/idea/symfony2plugin/tests/config/yaml/YamlCompletionContributorTest.java
@@ -307,116 +307,4 @@ public class YamlCompletionContributorTest extends SymfonyLightCodeInsightFixtur
             "$i"
         );
     }
-
-    public void testKeywordsCompletion() {
-        assertCompletionContains(YAMLFileType.YML, "" +
-                "services:\n" +
-                "    key: <caret>\n",
-            "true", ".inf"
-        );
-        assertCompletionContains(YAMLFileType.YML, "" +
-                "services:\n" +
-                "    key: <caret>",
-            "true", ".inf"
-        );
-        assertCompletionContains(YAMLFileType.YML, "" +
-                "services:\n" +
-                "    key: tr<caret>",
-            "true"
-        );
-//        assertCompletionContains(YAMLFileType.YML, "" +
-//                "services:\n" +
-//                "    key: .i<caret>",
-//            ".inf"
-//        );
-    }
-
-    public void testKeywordsCompletionInsideArray() {
-        assertCompletionContains(YAMLFileType.YML, "" +
-                "services:\n" +
-                "    key: [<caret>]\n",
-            "true", ".inf"
-        );
-        assertCompletionContains(YAMLFileType.YML, "" +
-                "services:\n" +
-                "    key: [FOO, <caret>]\n",
-            "true", ".inf"
-        );
-        assertCompletionContains(YAMLFileType.YML, "" +
-                "services:\n" +
-                "    key: [FOO, tr<caret>]\n",
-            "true"
-        );
-    }
-
-    public void testKeywordsCompletionInsideSequence() {
-        assertCompletionContains(YAMLFileType.YML, "" +
-                "services:\n" +
-                "    key:\n" +
-                "        - <caret>\n",
-            "true", ".inf"
-        );
-        assertCompletionContains(YAMLFileType.YML, "" +
-                "services:\n" +
-                "    key:\n" +
-                "        - tr<caret>\n",
-            "true"
-        );
-    }
-
-    public void testKeywordsCompletionInsideMapping() {
-        assertCompletionContains(YAMLFileType.YML, "" +
-                "services:\n" +
-                "    key: { foo: <caret> }\n",
-            "true", ".inf"
-        );
-        assertCompletionContains(YAMLFileType.YML, "" +
-                "services:\n" +
-                "    key: { foo: tr<caret> }\n",
-            "true"
-        );
-    }
-
-    public void testThatKeywordsAreNotCompletedAfterYamlTag() {
-        assertCompletionNotContains(YAMLFileType.YML, "" +
-                "services:\n" +
-                "    key: !mytag <caret>\n" +
-                "    foo: bar",
-            "true"
-        );
-        assertCompletionNotContains(YAMLFileType.YML, "" +
-                "services:\n" +
-                "    key: !mytag <caret>",
-            "true"
-        );
-    }
-
-    public void testThatKeywordsAreNotCompletedInNewLine() {
-        assertCompletionNotContains(YAMLFileType.YML, "" +
-                "services:\n" +
-                "    <caret>" +
-                "    foo: bar\n",
-            "true"
-        );
-        assertCompletionNotContains(YAMLFileType.YML, "" +
-                "services:\n" +
-                "    foo: bar\n" +
-                "<caret>",
-            "true"
-        );
-        assertCompletionNotContains(YAMLFileType.YML, "" +
-                "services:\n" +
-                "    foo: bar\n" +
-                "    <caret>",
-            "true"
-        );
-    }
-
-    public void testThatKeywordsAreNotCompletedInsideStringLiteral() {
-        assertCompletionNotContains(YAMLFileType.YML, "" +
-                "services:\n" +
-                "    foo: '<caret>'\n" ,
-            "true"
-        );
-    }
 }
